+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue #77: Add drop regularization
+- Issue #232:  Add supply sensor configuration
 
 ### Changed